Build Vagrant VM to test Ansible playbook
So far, my current laptop/desktop/localhost would be my testing ground.
This is terrible, because sometimes I would create config issues or
install something just for testing and would forget to remove.
On top of that I would setup my ansible playbook for a given
GNOME/Ansible/Manjaro/Arch version or package builds, only for that
config to fail whenever the versions or package build would change.

Being able to test against a Docker container or a Vagrant VM machine
should allow me to test the playbooks often and quicker, ensuring that
the setup is still valid.

This is the first step towards that goal.