Update Vagrant base image
Having to manage your own image is time consuming and I am struggling to
bring the size below 1GB.

Instead I will be using the Vagrant community to provide a base image
for this purpose.

The latest version of `Zelec/manjarolinux` is a base Manjaro image that
does not have Gnome installed.